+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Change log
 
+## in develop
+
+* Adds parameters for runtime and hint overrides to `Eval.Runtime` and `Eval.Meta` classes
+
 ## 0.15.0 (2021-08-16)
 
 * No longer unwraps optional type for an input with a default value